requirements.txt
streamlit
numpy
pandas
scikit-learn
opencv-python-headless
Pillow
joblib
scikit-image
matplotlib
tensorflow
